requirements.txt
Django
asgiref==3.7.2
psycopg2
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4

